// _db.types.ts
interface Story extends Document {
  _id: ObjectId;
  createdAt: string;
  updatedAt: string;
  initialized: boolean;
  title: string;
  story_title: string;
  public: boolean;
  summary: string;
  passage: string;
  story_art: {
    prompt: string;
    image_url: string;
  };
  owner: ObjectId;
  chapters: {
    title: string;
    tagline: string;
    chapter_number: number; // number of the chapter in the list
    audio_url: string; // keep this as an empty string
    content: { raw: string; markdown: string }; // Markdown and text String containing the content of the chapter.
    setup: {
      pov_character: string; // character whose perspective drives the chapter
      synopsis: string; // brief summary of the chapter's content
      scenes: {
        goal: string; // what needs to be accomplished in this scene
        setting: string; // where the scene takes place
        characters: string[]; // characters involved in the scene
        conflict: string; // Main tension or obstacle
        outcome: string; // resolution or consequence of the scene
        plot_points: string[]; // key plot points that need to be addressed in the scene
      }[];
    };
  }[];
// ...
}